Create NumberGuess.py
Number Guess

Wanna play a game? In this project, we’ll build a program that rolls a pair of dice and asks the user to guess the sum. If the user’s guess is equal to the total value of the dice roll, the user wins! Otherwise, the computer wins.

The program should do the following:
  Roll a pair of dice.
  Add the values of the roll.
  Ask the user to guess a number.
  Compare the user’s guess to the total value.
  Determine the winner (user or computer).